requirements.txt
amqp==5.0.6
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
billiard==3.6.4.0
celery==5.0.5
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.25.0
colorama==0.4.4
coverage==6.0.2
cryptography==3.4.8
defusedxml==0.7.1
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.5
django-allauth==0.45.0
django-appconf==1.0.4
django-celery-email==3.0.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-debug-toolbar==3.2
django-filter==2.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-meta==2.0.0
django-modelcluster==5.1
django-phonenumber-field==5.1.0
django-redis==4.12.1
django-robots==4.0
django-taggit==1.5.1
django-templated-mail==1.1.1
django-treebeard==4.5.1
Faker==8.11.0
filelock==3.0.12
flake8==3.9.2
gunicorn==20.1.0
idna==3.2
iniconfig==1.1.1
kombu==5.1.0
mccabe==0.6.1
oauthlib==3.1.1
packaging==21.0
phonenumbers==8.12.20
Pillow==8.3.1
platformdirs==2.2.0
pluggy==0.13.1
prompt-toolkit==3.0.19
psycopg2==2.9.1
psycopg2-binary==2.9.1
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
python-dateutil==2.8.2
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
sentry-sdk==1.4.3
six==1.16.0
sqlparse==0.4.1
stripe==2.60.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.6
vine==5.0.0
virtualenv==20.6.0
wcwidth==0.2.5
whitenoise==5.3.0
